requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
chainlit==2.9.4
requests==2.32.5
# jwt==1.4.0
PyJWT[crypto]==2.8.0
cryptography>=41
itsdangerous==2.2.0
sqlalchemy==2.0.45
pyodbc==5.3.0
azure-monitor-opentelemetry==1.8.3